Banish foul demons from the dungeons of event system
git-svn-id: https://skeldal.svn.sourceforge.net/svnroot/skeldal@70 27359af5-ae43-0410-a23f-8b0a9d901db5

+++ branches/next_ghost/sdl/input.c
@@ -63,7 +63,7 @@ int Input_ReadKey(void) {
 
 void Sys_ProcessEvents(void) {
 	SDL_Event event;
-	uint16_t keycode;
+	int keycode;
 
 	while (SDL_PollEvent(&event)) {
 		switch (event.type) {
